+++ test/com/mobiletheatertech/plot/HangPointTest.java
@@ -39,8 +39,8 @@ public class HangPointTest {
         HangPoint hangPoint = new HangPoint( element );
 
         assertEquals( TestHelpers.accessString( hangPoint, "id" ), "Blather" );
-        assertEquals( TestHelpers.accessInteger( hangPoint, "x" ), 296 );
-        assertEquals( TestHelpers.accessInteger( hangPoint, "y" ), 320 );
+        assertEquals( TestHelpers.accessInteger( hangPoint, "x" ), (Integer) 296 );
+        assertEquals( TestHelpers.accessInteger( hangPoint, "y" ), (Integer) 320 );
     }
 
     @Test
@@ -116,63 +116,63 @@ public class HangPointTest {
         new HangPoint( element );
     }
 
-    @Test( expectedExceptions = AttributeMissingException.class,
-           expectedExceptionsMessageRegExp = "HangPoint \\(Blather\\) is missing required 'x' attribute." )
+    @Test(expectedExceptions = AttributeMissingException.class,
+          expectedExceptionsMessageRegExp = "HangPoint \\(Blather\\) is missing required 'x' attribute.")
     public void noX() throws Exception {
         element.removeAttribute( "x" );
         new HangPoint( element );
     }
 
-    @Test( expectedExceptions = AttributeMissingException.class,
-           expectedExceptionsMessageRegExp = "HangPoint instance is missing required 'x' attribute." )
+    @Test(expectedExceptions = AttributeMissingException.class,
+          expectedExceptionsMessageRegExp = "HangPoint instance is missing required 'x' attribute.")
     public void noXWithoutID() throws Exception {
         element.removeAttribute( "id" );
         element.removeAttribute( "x" );
         new HangPoint( element );
     }
 
-    @Test( expectedExceptions = AttributeMissingException.class,
-           expectedExceptionsMessageRegExp = "HangPoint \\(Blather\\) is missing required 'y' attribute." )
+    @Test(expectedExceptions = AttributeMissingException.class,
+          expectedExceptionsMessageRegExp = "HangPoint \\(Blather\\) is missing required 'y' attribute.")
     public void noY() throws Exception {
         element.removeAttribute( "y" );
         new HangPoint( element );
     }
 
-    @Test( expectedExceptions = AttributeMissingException.class,
-           expectedExceptionsMessageRegExp = "HangPoint instance is missing required 'y' attribute." )
+    @Test(expectedExceptions = AttributeMissingException.class,
+          expectedExceptionsMessageRegExp = "HangPoint instance is missing required 'y' attribute.")
     public void noYWithoutID() throws Exception {
         element.removeAttribute( "id" );
         element.removeAttribute( "y" );
         new HangPoint( element );
     }
 
-    @Test( expectedExceptions = LocationException.class,
-           expectedExceptionsMessageRegExp =
-                   "HangPoint x value outside boundary of the venue" )
+    @Test(expectedExceptions = LocationException.class,
+          expectedExceptionsMessageRegExp =
+                  "HangPoint x value outside boundary of the venue")
     public void tooLargeX() throws Exception {
         element.setAttribute( "x", "351" );
         new HangPoint( element );
     }
 
-    @Test( expectedExceptions = LocationException.class,
-           expectedExceptionsMessageRegExp =
-                   "HangPoint y value outside boundary of the venue" )
+    @Test(expectedExceptions = LocationException.class,
+          expectedExceptionsMessageRegExp =
+                  "HangPoint y value outside boundary of the venue")
     public void tooLargeY() throws Exception {
         element.setAttribute( "y", "401" );
         new HangPoint( element );
     }
 
-    @Test( expectedExceptions = LocationException.class,
-           expectedExceptionsMessageRegExp =
-                   "HangPoint x value outside boundary of the venue" )
+    @Test(expectedExceptions = LocationException.class,
+          expectedExceptionsMessageRegExp =
+                  "HangPoint x value outside boundary of the venue")
     public void tooSmallX() throws Exception {
         element.setAttribute( "x", "-1" );
         new HangPoint( element );
     }
 
-    @Test( expectedExceptions = LocationException.class,
-           expectedExceptionsMessageRegExp =
-                   "HangPoint y value outside boundary of the venue" )
+    @Test(expectedExceptions = LocationException.class,
+          expectedExceptionsMessageRegExp =
+                  "HangPoint y value outside boundary of the venue")
     public void tooSmallY() throws Exception {
         element.setAttribute( "y", "-1" );
         new HangPoint( element );
