+++ test/com/mobiletheatertech/plot/PipeTest.java
@@ -128,7 +128,7 @@ public class PipeTest {
     public void storesAttributes() throws Exception {
         Pipe pipe = new Pipe( element );
 
-        assertEquals( TestHelpers.accessInteger( pipe, "length" ), 120 );
+        assertEquals( TestHelpers.accessInteger( pipe, "length" ), (Integer) 120 );
         assertTrue( new Point( 12, 23, 34 ).equals( TestHelpers.accessPoint( pipe, "start" ) ) );
         assertEquals( TestHelpers.accessPoint( pipe, "start" ), new Point( 12, 23, 34 ) );
         assertEquals( TestHelpers.accessString( pipe, "id" ), "" );
